move membersListAsResourceData into src/lib/util/
