+++ src/lib/operations/readContainer.ts
@@ -1,7 +1,7 @@
 import Debug from 'debug'
 import { WacLdpResponse, ResultType } from '../api/http/HttpResponder'
 import { WacLdpTask } from '../api/http/HttpParser'
-import { membersListAsResourceData } from '../../membersListAsResourceData'
+import { membersListAsResourceData } from '../util/membersListAsResourceData'
 import { BlobTree } from '../storage/BlobTree'
 
 const debug = Debug('readContainer')
